+++ CHANGELOG.md
@@ -7,6 +7,10 @@
   useful summary for people upgrading their application, not a replication
   of the commit log.
 
+## Unreleased
+
+* Allow GA4 link tracker to track to multiple child classes ([PR #3835](https://github.com/alphagov/govuk_publishing_components/pull/3835))
+
 ## 37.2.4
 
 * Fix analytics consent ([PR #3833](https://github.com/alphagov/govuk_publishing_components/pull/3833))